Move RawReadResponse into common; remove into_read_responses
