+++ downstairs/src/extent_inner_raw.rs
@@ -7,11 +7,10 @@ use crate::{
     },
     integrity_hash, mkdir_for_file,
     region::{BatchedPwritev, JobOrReconciliationId},
-    Block, BlockContext, CrucibleError, JobId, RawReadResponse,
-    RegionDefinition,
+    Block, BlockContext, CrucibleError, JobId, RegionDefinition,
 };
 
-use crucible_protocol::ReadResponseBlockMetadata;
+use crucible_protocol::{RawReadResponse, ReadResponseBlockMetadata};
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 use slog::{error, Logger};
@@ -1473,7 +1472,7 @@ mod test {
     use anyhow::Result;
     use bytes::{Bytes, BytesMut};
     use crucible_protocol::EncryptionContext;
-    use crucible_protocol::{ReadRequest, ReadResponse};
+    use crucible_protocol::ReadRequest;
     use rand::Rng;
     use tempfile::tempdir;
 
@@ -1733,14 +1732,14 @@ mod test {
 
             // We should not get back our data, because block 0 was written.
             assert_ne!(
-                resp,
-                vec![ReadResponse {
+                resp.blocks,
+                vec![ReadResponseBlockMetadata {
                     eid: 0,
                     offset: Block::new_512(0),
-                    data: BytesMut::from(data.as_ref()),
                     block_contexts: vec![block_context]
                 }]
             );
+            assert_ne!(resp.data, BytesMut::from(data.as_ref()));
         }
 
         // But, writing to the second block still should!
@@ -1767,14 +1766,14 @@ mod test {
 
             // We should get back our data! Block 1 was never written.
             assert_eq!(
-                resp,
-                vec![ReadResponse {
+                resp.blocks,
+                vec![ReadResponseBlockMetadata {
                     eid: 0,
                     offset: Block::new_512(1),
-                    data: BytesMut::from(data.as_ref()),
                     block_contexts: vec![block_context]
                 }]
             );
+            assert_eq!(resp.data, BytesMut::from(data.as_ref()));
         }
 
         Ok(())
@@ -1990,14 +1989,14 @@ mod test {
 
             // We should get back our data! Block 1 was never written.
             assert_eq!(
-                resp,
-                vec![ReadResponse {
+                resp.blocks,
+                vec![ReadResponseBlockMetadata {
                     eid: 0,
                     offset: Block::new_512(0),
-                    data: BytesMut::from(data.as_ref()),
                     block_contexts: vec![block_context]
                 }]
             );
+            assert_eq!(resp.data, BytesMut::from(data.as_ref()));
         }
 
         Ok(())
@@ -2501,15 +2500,15 @@ mod test {
         };
 
         assert_eq!(
-            resp,
-            vec![ReadResponse {
+            resp.blocks,
+            vec![ReadResponseBlockMetadata {
                 eid: 0,
                 offset: Block::new_512(0),
-                data: BytesMut::from(data.as_ref()),
                 // Only the most recent block context should be returned
                 block_contexts: vec![block_context],
             }]
         );
+        assert_eq!(resp.data, BytesMut::from(data.as_ref()));
 
         Ok(())
     }
